backend/pyproject.toml
[project]
name = "email-marketing-backend"
version = "0.1.0"
description = "Flask service for the enterprise email marketing platform"
authors = [{ name = "Email Marketing Team" }]
requires-python = ">=3.11"
dependencies = [
    "Flask>=3.1.0",
    "Flask-Cors>=5.0.0",
    "Flask-Migrate>=4.0.7",
    "Flask-SQLAlchemy>=3.1.1",
    "SQLAlchemy>=2.0.38",
    "email-validator>=2.2.0",
    "celery>=5.4.0",
    "redis>=5.2.1",
    "psycopg[binary]>=3.2.3",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "gunicorn>=23.0.0",
    "PyJWT>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
    "ruff>=0.7.3",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]
extend-select = [
    "I",
]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
